+++ components/MetadataControl/samples/MetadataControl.md
@@ -13,14 +13,13 @@ issue-id: 0
 
 # MetadataControl
 
-The [MetadataControl](/dotnet/api/microsoft.toolkit.uwp.ui.controls.metadatacontrol) control displays a
-list of labels and hyper-links separated by a bullet.
+The [MetadataControl](/dotnet/api/microsoft.toolkit.uwp.ui.controls.metadatacontrol) control displays a list of labels and hyper-links separated by a bullet.
 It also generates an accessible string representing its content.
 
 The bullet separator can be customized using the `Separator` property.
 `AccessibleSeparator` is used as a replacement for `Separator` to generate the accessible string.
 
-The control needs a list of [MediadataItem](https://github.com/windows-toolkit/WindowsCommunityToolkit/blob/main/Microsoft.Toolkit.Uwp.UI.Controls.Core/MetadataControl/MetadataItem.cs).
+The control needs a list of [MetadataItem](https://github.com/windows-toolkit/WindowsCommunityToolkit/blob/main/Microsoft.Toolkit.Uwp.UI.Controls.Core/MetadataControl/MetadataItem.cs).
 Each item will be displayed either as a text or as an hyper-link (if the `Command`property is set).
 
 The default control template is using on a `TextBlock`. The style of this `TextBlock` can be customized using the `TextBlockStyle` property.
@@ -48,9 +47,9 @@ metadataControl.Items = new[]
 };
 ```
 
-## MediadataItem
+## MetadataItem
 
-A `MediadataItem` contains the information about one entry which will be displayed in the `MetadataControl`
+A `MetadataItem` contains the information about one entry which will be displayed in the `MetadataControl`
 
 | Property | Type | Description |
 | -- | -- | -- |
